refactor : separate random power generator
