+++ src/main/java/racingcar/view/IO.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.model.Car;
 
 import java.util.ArrayList;
 import java.util.Arrays;
