+++ src/test/java/racingcar/CarTest.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.model.Car;
 
 import static org.junit.jupiter.api.Assertions.*;
